Rename /src/utils/resolve-url -> utils
